Actualiza `pubspec.yaml` para reemplazar imágenes de banner de formato JPG a PNG. Se eliminan los archivos JPG correspondientes y se modifica `home_layout.dart` para utilizar el nuevo componente `HeaderCarousel`, simplificando la gestión del carrusel de imágenes.
